requirements.txt
discord==2.1.0
google-api-python-client==2.78.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0